增加rust调试文档

+++ scripts/fivestar/miner.sh
@@ -30,5 +30,7 @@ else
 fi
 
 wait "$pid"
-kill $pid
+if [ ! -z "$pid"]; then
+    kill $pid
+fi
